Data on box office receipts and viewings
Data on collections and views of national Russian films. Information from open sources for the period from January 2004 to December 2024. The material is permitted for use in scientific and research purposes, as well as for the implementation of educational programs without any restrictions with the obligatory indication of the link to the source and authorship